+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.0.0] - 2026-01-20
+
+### Changed
+
+- Remove flake8 and black tasks
+- Remove the `*-new` tasks now that ruff has been implimented in all repos
+
 ## [1.7.1] - 2026-01-20
 
 ### Fixed